remove code generation, api crds sync from vendor
Previously we were doing our own deepcopy gen for our API in the MCO
itself, but now that we've migrated it, we shouldn't do that anymore.

Also, since our CRDs are generated in openshift/api now, we need to
update them out there, and then vendor them back in. This adds a
hack/crds-sync.sh script that will copy the CRDs from vendor, and also
updates the Makefile such that this CRD update happens when you run
"make update".