update kubelet templates with new api featuregates
We migrated our api into a newer version of openshift/api than we
initially had in the MCO, so when we bumped the api version to bring our
own API back in, we brought some new feature gates with it. This just
adds those new default featuregates to the kubelet template so
everything is in sync and the test that checks this will pass.